Updated GLFW to 2.7.8 and resolved some minor problems. This version is tested in the dome.

+++ src/sgct/SharedData.cpp
@@ -522,7 +522,7 @@ unsigned char * SharedData::readUCharArray(size_t length)
     Engine::lockMutex(sgct_core::NetworkManager::gMutex);
 
 	unsigned char * p = &dataBlock[pos];
-    pos += length;
+    pos += static_cast<unsigned int>(length);
 
 	Engine::unlockMutex(sgct_core::NetworkManager::gMutex);
